requirements.txt
asgiref==3.3.1
defusedxml==0.6.0
diff-match-patch==20200713
dj-static==0.0.6
Django==3.1.3
django-ckeditor==6.0.0
django-import-export==2.4.0
django-js-asset==1.2.2
et-xmlfile==1.0.1
gunicorn==20.0.4
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
psycopg2==2.8.6
pytz==2020.4
PyYAML==5.3.1
sqlparse==0.4.1
static3==0.7.0
tablib==2.0.0
xlrd==1.2.0
xlwt==1.3.0
